+++ Assets/Scripts/ProjectileData.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class ProjectileData : ScriptableObject
 {
     public GameObject projectilePrefab;
-    public int projectileDamage;
+    public float projectileDamage;
     public float projectileSpeed;
     public float projectileRange;
     public float projectileFireRate;
